SCREENER_MIN_DAYS_TO_EXPIRY 类型 int 改 float：支持 0.5（半天）等小数到期阈值

+++ guardian_v8/config.py
@@ -109,7 +109,7 @@ class Config:
         )
     )
     screener_min_daily_rewards: float = float(os.environ.get("SCREENER_MIN_DAILY_REWARDS", "10.0"))
-    screener_min_days_to_expiry: int = int(os.environ.get("SCREENER_MIN_DAYS_TO_EXPIRY", "0"))
+    screener_min_days_to_expiry: float = float(os.environ.get("SCREENER_MIN_DAYS_TO_EXPIRY", "0"))
     screener_min_midpoint: float = float(os.environ.get("SCREENER_MIN_MIDPOINT", "0.15"))
     screener_max_midpoint: float = float(os.environ.get("SCREENER_MAX_MIDPOINT", "0.85"))
     screener_min_size_lower: float = float(os.environ.get("SCREENER_MIN_SIZE_LOWER", "0.0"))
